Continue refactor.
