--- .github/workflows/format.yml
@@ -1,15 +1,0 @@
-name: Format Check
-
-on: [push, pull_request]
-
-jobs:
-  format:
-    name: Format Check
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v4
-    - name: Run clang-format style check for C/C++/Protobuf programs.
-      uses: jidicula/clang-format-action@v4.18.0
-      with:
-        clang-format-version: '22'
-        check-path: '.'
