+++ common/include/Eigen.hh
@@ -14,15 +14,16 @@ concept HasDimension = requires {
 };
 
 template <typename T>
-concept SupportsVectorOperations =
-    HasDimension<T> &&
-    requires(T a, T b, double scalar,
-             const Eigen::Vector<double, T::Dimension>& vec) {
-      { a + b } -> std::convertible_to<T>;
-      { a - b } -> std::convertible_to<T>;
-      { a + vec } -> std::convertible_to<T>;
-      { a - vec } -> std::convertible_to<T>;
-    };
+concept SupportsVectorOperations = requires(
+    T a, T b, double scalar, const Eigen::Vector<double, T::Dimension>& vec) {
+  { a + b } -> std::convertible_to<T>;
+  { a - b } -> std::convertible_to<T>;
+  { a + vec } -> std::convertible_to<T>;
+  { a - vec } -> std::convertible_to<T>;
+};
+
+template <typename T>
+concept Vector = HasDimension<T> && SupportsVectorOperations<T>;
 
 template <typename Derived, int Dim>
 class VectorBase {
@@ -103,33 +104,9 @@ template <int States, int Inputs>
 using InputLeftPseudoInverseMatrix = Eigen::Matrix<double, Inputs, States>;
 
 // Convenience aliases for commonly-used matrix/matrix-pair shapes
-// NOTE(hayden): "Matrices" is ambiguous
+// TODO(hayden): "Matrices" is ambiguous
 template <int States, int Inputs>
 using Matrices = std::pair<SystemMatrix<States>, InputMatrix<States, Inputs>>;
-
-// --- Time-derivative / dot aliases ---
-//
-// Provide a trait-based mechanism so specific State types can map to
-// canonical derivative wrapper types (e.g., PositionAccelerationState) while
-// allowing a sensible default. This avoids attempting to explicitly specialize
-// alias templates (which the language forbids).
-//
-// Default behaviour: TimeDerivative<State> == State
-// To customize for a particular State type, specialize TimeDerivativeOf<State>
-// (for example, in `state.hh`) like:
-//
-//   template <>
-//   struct TimeDerivativeOf<PositionVelocityState> {
-//     using type = PositionAccelerationState;
-//   };
-//
-// Aliases:
-//   TimeDerivative<State> -> typename TimeDerivativeOf<State>::type
-//   Dot<State>             -> TimeDerivative<State>
-//
-// NumericTimeDerivative<State> remains the raw Eigen vector type
-// (Eigen::Vector<double, State::Dimension>) for cases that need the plain
-// numeric representation rather than a wrapper.
 
 template <typename State>
 struct TimeDerivativeOf {
@@ -139,23 +116,22 @@ struct TimeDerivativeOf {
 template <typename State>
 using TimeDerivative = typename TimeDerivativeOf<State>::type;
 
+// Define `Dot` as an alias for `TimeDerivative` so library users can write
+// `AccelerationVector = Dot<VelocityVector>`
 template <typename State>
 using Dot = TimeDerivative<State>;
 
-// Explicit numeric/raw time-derivative when you need the plain Eigen vector.
-template <typename State>
-using NumericTimeDerivative = Eigen::Vector<double, State::Dimension>;
-
+// Discretizes a pair of continuous system-input matrices for the sample period
 template <int States, int Inputs>
-Matrices<States, Inputs> Discretize(Matrices<States, Inputs>& AcBc,
+Matrices<States, Inputs> Discretize(Matrices<States, Inputs>& Ac_Bc,
                                     quantities::Time sample_period) {
   using BlockMatrix = Eigen::Matrix<double, States + Inputs, States + Inputs>;
 
   // M = ⎡ Ac Bc ⎤
   //     ⎣ 0  0  ⎦
   BlockMatrix M;
-  M.template block<States, States>(0, 0) = AcBc.first;
-  M.template block<States, Inputs>(0, States) = AcBc.second;
+  M.template block<States, States>(0, 0) = Ac_Bc.first;
+  M.template block<States, Inputs>(0, States) = Ac_Bc.second;
   M.template block<Inputs, States + Inputs>(States, 0).setZero();
 
   // ϕ = ⎡ Ad Bd ⎤
@@ -167,6 +143,7 @@ Matrices<States, Inputs> Discretize(Matrices<States, Inputs>& AcBc,
   return std::make_pair(Ad, Bd);
 }
 
+// Calculates the left pseudo-inverse for the continuous input matrix
 template <int States, int Inputs>
 InputLeftPseudoInverseMatrix<States, Inputs> PseudoInverse(
     const InputMatrix<States, Inputs>& Bc) {
@@ -175,15 +152,12 @@ InputLeftPseudoInverseMatrix<States, Inputs> PseudoInverse(
   return (BcT * Bc).inverse() * BcT;
 }
 
-template <int Inputs, int States, int Gains>
-  requires(Gains == States)
+// Constructs a gain matrix mapping from states vectors to input vectors.
+// Simplifies constructing feedback controllers, where state vectors represent
+// error between a reference and current state.
+template <int Inputs, int States>
 Eigen::Matrix<double, Inputs, States> MakeGainMatrix(
-    const Eigen::Matrix<double, Gains, 1>& gains) {
-  // Gain matrix K should map state (States x 1) -> input (Inputs x 1):
-  //   input = K * state
-  // so K must be (Inputs x States). Populate each column i with the scalar
-  // gain `gains[i]`. For Inputs>1 this replicates the scalar across the input
-  // channels for that state dimension (consistent fallback).
+    const Eigen::Matrix<double, States, 1>& gains) {
   Eigen::Matrix<double, Inputs, States> result =
       Eigen::Matrix<double, Inputs, States>::Zero();
   for (int i = 0; i < States; ++i) {
