+++ common/include/robot.hh
@@ -5,7 +5,7 @@
 #include "au/math.hh"
 #include "au/units/inches.hh"
 #include "au/units/pounds_mass.hh"
-#include "system/Motor.hh"
+#include "system/motor/Motor.hh"
 #include "units.hh"
 
 namespace reefscape {
